[#1] Project skeleton: small refactoring (move controllers to module)
