+++ src/templates/base.html
@@ -5,11 +5,11 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Podcasts - Home</title>
-    <script src="/src/static/js/tailwind.3.4.17.js"></script>
-    <script src="/src/static/js/colors.min.js"></script>
-    <script src="/src/static/js/flatten-color-palette.min.js"></script>
-    <script src="/src/static/js/default-theme.min.js"></script>
-    <script src="/src/static/js/plugin.min.js"></script>
+    <script src="/static/js/tailwind.3.4.17.js"></script>
+    <script src="/static/js/colors.min.js"></script>
+    <script src="/static/js/flatten-color-palette.min.js"></script>
+    <script src="/static/js/default-theme.min.js"></script>
+    <script src="/static/js/plugin.min.js"></script>
     {% endblock %}
 </head>
 <body class="flex min-h-screen">
