+++ src/templates/index.html
@@ -14,7 +14,7 @@
             </div>
         {% endfor %}
         <div class="max-w-[250px] rounded overflow-hidden shadow-lg">
-            <img class="w-full" src="/src/static/images/podcast-add.png" alt="Podcast Add">
+            <img class="w-full" src="/static/images/podcast-add.png" alt="Podcast Add">
             <div class="px-6 py-4">
                 <div class="font-bold text-xl mb-2">🆕 Add new one!</div>
                 <p class="text-gray-700 text-base">
